+++ lib/providers/node_provider/state/state_manager_interface.dart
